pyproject.toml
[project]
name = "git-tool"
version = "0.1.0"
description = ""
authors = [
    {name = "S29Ducky",email = "nbarasraso@edenschool.fr"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "pygithub (>=2.8.1,<3.0.0)",
    "rich (>=14.2.0,<15.0.0)",
    "prompt-toolkit (>=3.0.52,<4.0.0)"
]

[tool.poetry]
packages = [{include = "git_tool", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
